OpenGLWrapper_JE/requirements.txt
glfw==2.1.0
Pillow==9.0.0
PyOpenGL==3.1.5
